smartrag/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
kafka-python==2.0.2
email-validator==2.1.0
pydantic-core==2.14.1
loguru==0.7.2
httpx==0.25.2
openai==1.3.7
python-docx==0.8.11
PyPDF2==3.0.1
Pillow==10.1.0